requirements.txt
Django==4.1.4
djangorestframework==3.14.0
dj-rest-auth==3.0.0
django-import-export==3.2.0
django-crispy-forms==1.11.2
gunicorn==20.1.0
pytz
whitenoise

